fix(subnet-sdk): Go emitted base64 where the constructor wanted level arrays
`MeshConfig.SubnetAttachment` and `SubnetPath.Levels` were `[]uint8`.
`[]uint8` IS `[]byte` in Go, and `encoding/json` special-cases that as
BASE64 on the way out — so a provider's attachment reached `net_mesh_new`
as `"Awk="` instead of `[3,9]`, and the node refused the entire config
with `invalid JSON`. Every Go provider configured with a subnet
attachment or a named export was affected: P1-7 and P1-6 shipped a
configuration surface that could not construct a node.

Both fields are now `[]uint32`, which is what `MeshConfig.Subnet` has
always been — the codebase already knew this and I broke the pattern.

The asymmetry is why it shipped: unmarshalling `[3,9]` INTO `[]uint8`
SUCCEEDS. Only marshalling is base64. So the live cell's manifest parsed
cleanly and the failure surfaced one call later at construction, looking
like a bad config rather than a bad type. Verified both directions with
a standalone probe before changing anything.

`DeclareSubnetBoundaries` now range-guards each level instead of casting
to `uint8`: the JSON-facing type is wider than a path label, and silently
wrapping 256 to 0 would move the declared boundary.

Regression pinned on the Rust side, where it can run on any host: a
captured sample of `json.Marshal(MeshConfig{...})` must deserialize into
`MeshNewConfig`, and the base64 shape a `[]uint8` field would produce
must be REFUSED — so the first test cannot pass by the deserializer
being lenient. The Go live cell remains the guard for a Go-side type
regression; it runs in CI, where `go test ./...` auto-discovers it.

Not typecheckable here: `go vet` cannot resolve this cgo package on this
host at all (it loses `MeshNode` itself, though `go build ./...`
succeeds), so the Go changes are inspection + the pinned JSON contract.

+++ go/mesh.go
@@ -150,7 +150,10 @@ type MeshConfig struct {
 	// levels each 0–255. Distinct from Subnet; omitting it preserves the
 	// core compatibility fallback, which protected deployments should not
 	// rely on.
-	SubnetAttachment []uint8 `json:"subnet_attachment,omitempty"`
+	//
+	// `[]uint32` for the same reason as Subnet above and SubnetPath.Levels:
+	// encoding/json emits `[]uint8` as base64, which the constructor rejects.
+	SubnetAttachment []uint32 `json:"subnet_attachment,omitempty"`
 
 	// SubnetControlChannel treats an ordinary configured channel as a subnet
 	// control-fact ARRIVAL path. Confers no authority — facts verify by
